ajustments on docker compose for app running
