+++ .env
@@ -1,0 +1,5 @@
+DB_USER=postgres
+DB_PASSWORD=postgres
+DB_NAME=orch
+DB_PORT=5432
+PORT=3030
