+++ .gitignore
@@ -19,5 +19,3 @@
 
 # Go workspace file
 go.work
-
-.env
