+++ Dockerfile
@@ -3,13 +3,12 @@ FROM golang:1.20
 
 WORKDIR /app
 
-COPY go.mod ./
-COPY main.go ./
+COPY . .
 
 RUN go mod tidy
 
-RUN go build -o main.go 
+RUN go build -o api ./main.go
 
 EXPOSE 3000
 
-CMD [ "./main.go" ]
+CMD [ "./api" ]
