+++ docker-compose.yml
@@ -1,11 +1,13 @@
 version: '3.8'
 
 services:
-  web:
+  api:
     container_name: api
-    build: .
+    build:
+      context: .
+      dockerfile: ./Dockerfile
     env_file:
-      - .env.local
+      - .env
     ports:
       - 3000:3000
     depends_on:
@@ -16,12 +18,7 @@ services:
     ports:
       - 5432:5432
     environment:
-      - POSTGRES_USER=${DB_USER}
-      - POSTGRES_PASSWORD=${DB_PASSWORD}
-      - POSTGRES_DB=${DB_NAME}
-
-    volumes:
-      - postgres-db:/var/lib/postgresql/data
-
-volumes:
-  postgres-db:
+      POSTGRES_USER: ${DB_USER}
+      POSTGRES_PASSWORD: ${DB_PASSWORD}
+      POSTGRES_DB: ${DB_NAME}
+      PGPORT: ${DB_PORT}
